+++ packages/session-replay-browser/src/config/local-config.ts
@@ -9,6 +9,7 @@ import {
   SessionReplayVersion,
 } from './types';
 import { SafeLoggerProvider } from '../logger';
+import { validateUGCFilterRules } from '../helpers';
 
 export const getDefaultConfig = () => ({
   flushMaxRetries: 2,
@@ -30,6 +31,7 @@ export class SessionReplayLocalConfig extends Config implements ISessionReplayLo
   storeType: StoreType;
   performanceConfig?: SessionReplayPerformanceConfig;
   experimental?: { useWebWorker: boolean };
+  applyBackgroundColorToBlockedElements?: boolean;
 
   constructor(apiKey: string, options: SessionReplayOptions) {
     const defaultConfig = getDefaultConfig();
@@ -53,9 +55,18 @@ export class SessionReplayLocalConfig extends Config implements ISessionReplayLo
     this.version = options.version;
     this.performanceConfig = options.performanceConfig;
     this.storeType = options.storeType ?? 'idb';
+    this.applyBackgroundColorToBlockedElements = options.applyBackgroundColorToBlockedElements ?? false;
 
     if (options.privacyConfig) {
       this.privacyConfig = options.privacyConfig;
+    }
+    if (options.interactionConfig) {
+      this.interactionConfig = options.interactionConfig;
+
+      // validate ugcFilterRules, throw error if invalid - throw error at the beginning of the config
+      if (this.interactionConfig.ugcFilterRules) {
+        validateUGCFilterRules(this.interactionConfig.ugcFilterRules);
+      }
     }
     if (options.debugMode) {
       this.debugMode = options.debugMode;